Session 7 - FAISS-and-Advanced-RAG/pyproject.toml
[project]
name = "eag-v1-2025-Session-7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "google-genai>=1.25.0",
    "python-dotenv",
    "mcp[cli]",
    "Pillow",
    "pywinauto",
    "faiss-cpu",
    "requests",
    "ollama",
    "scipy>=1.16.3",
    "tqdm>=4.67.1",
    "markitdown[all]>=0.1.3",
]
